Move checking camera permission to CameraVideoCapturer
